Декомпозирует RoomScreen
