+++ project/src/index.tsx
@@ -3,7 +3,6 @@ import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
 
 import { store } from './store';
-import { fetchOffersAction } from './store/offers/api-actions';
 import { checkAuthAction } from './store/user/api-actions';
 
 import App from './components/app/app';
@@ -12,7 +11,6 @@ const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement,
 );
 
-store.dispatch(fetchOffersAction());
 store.dispatch(checkAuthAction());
 
 root.render(
