+++ project/src/pages/main-screen/main-screen.tsx
@@ -2,10 +2,14 @@ import { useEffect } from 'react';
 
 import { useAppDispatch, useAppSelector } from '../../hooks';
 
+import { store } from '../../store';
 import { updateLayout } from '../../store/layout/actions';
+import { fetchOffersAction } from '../../store/offers/api-actions';
 
 import Loader from '../../components/loader/loader';
 import MainScreenContent from '../../components/main-screen-content/main-screen-content';
+
+store.dispatch(fetchOffersAction());
 
 function MainScreen() {
   const areOffersLoading = useAppSelector((state) => state.offers.areOffersLoading);
